+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVC1VC0F2RH0ZEYT7GP9KW3P"
+        "01JVC8YGP3253J8QK9V1BJB3SB"
     )
     r"""The id of the resource"""