+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-16T08:02:19.729Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-16T10:06:22.724Z")
     r"""The date and time the resource was last updated"""