+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVC1VC0EWNPEB2KHP6YV0S7X"
+        "01JVC8YGP2BJSGFWHXXWKWRQ13"
     )
     r"""The id of the resource"""